separate out unencrypted store file handler
