+++ pod_store/store.py
@@ -1,112 +1,11 @@
-import json
 import os
+
 from typing import List, Optional
 
 from .exc import PodcastDoesNotExistError, PodcastExistsError, StoreExistsError
 from .podcasts import Podcast
+from .store_file_handlers import StoreFileHandler, UnencryptedStoreFileHandler
 from .util import run_git_command
-
-
-class StoreFileHandler:
-    """Class for reading/writing data from the store file.
-
-    _store_file_path (str): file system location of the json file that holds store data.
-    """
-
-    def __init__(self, store_file_path):
-        self._store_file_path = store_file_path
-
-    @classmethod
-    def create_with_file(cls, store_file_path: str):
-        """Creates an empty store file while constructing the class."""
-        file_handler = cls(store_file_path)
-        file_handler.write_data({})
-        return file_handler
-
-    def __repr__(self):
-        return "<StoreFileHandler({self._store_file_path!r})>"
-
-    def read_data(self):
-        """Return json data from the store file."""
-        with open(self._store_file_path) as f:
-            return json.load(f)
-
-    def write_data(self, data: dict):
-        """Write json data to the store file."""
-        with open(self._store_file_path, "w") as f:
-            json.dump(data, f, indent=2)
-
-
-class Store:
-    """Podcast store coordinating class.
-
-    podcasts (StorePodcasts): tracks all the podcasts kept in the store
-
-    _store_path (str): location of pod store directory
-
-    _podcast_downloads_path (str): file system directory for podcasts to download
-        episodes into
-
-    _file_handler (StoreFileHandler): class that handles reading/writing from the store
-        json file
-    """
-
-    def __init__(
-        self,
-        store_path: str,
-        podcast_downloads_path: str,
-        file_handler: StoreFileHandler,
-    ) -> None:
-        self._store_path = store_path
-        self._podcast_downloads_path = podcast_downloads_path
-
-        self._file_handler = file_handler
-
-        podcast_data = file_handler.read_data()
-        self.podcasts = StorePodcasts(
-            podcast_data=podcast_data, podcast_downloads_path=podcast_downloads_path
-        )
-
-    @classmethod
-    def create(
-        cls,
-        store_path: str,
-        store_file_path: str,
-        podcast_downloads_path: str,
-        setup_git: bool,
-        git_url: Optional[str] = None,
-        store_file_handler_cls: StoreFileHandler = StoreFileHandler,
-    ):
-        """Initialize a new pod store.
-
-        Optionally set up the `git` repo for the store.
-        """
-        try:
-            os.makedirs(store_path)
-        except FileExistsError:
-            raise StoreExistsError(store_path)
-        os.makedirs(podcast_downloads_path, exist_ok=True)
-
-        file_handler = store_file_handler_cls.create_with_file(store_file_path)
-
-        if setup_git:
-            run_git_command("init")
-            if git_url:
-                run_git_command(f"remote add origin {git_url}")
-
-        return cls(
-            store_path=store_path,
-            podcast_downloads_path=podcast_downloads_path,
-            file_handler=file_handler,
-        )
-
-    def __repr__(self):
-        return f"<Store({self._store_path!r})>"
-
-    def save(self):
-        """Save data to the store json file."""
-        podcast_data = self.podcasts.to_json()
-        self._file_handler.write_data(podcast_data)
 
 
 class StorePodcasts:
@@ -213,3 +112,75 @@ class StorePodcasts:
     def to_json(self):
         """Convert store podcasts to json data for writing to the store file."""
         return {title: podcast.to_json() for title, podcast in self._podcasts.items()}
+
+
+class Store:
+    """Podcast store coordinating class.
+
+    podcasts (StorePodcasts): tracks all the podcasts kept in the store
+
+    _store_path (str): location of pod store directory
+
+    _podcast_downloads_path (str): file system directory for podcasts to download
+        episodes into
+
+    _file_handler (StoreFileHandler): class that handles reading/writing from the store
+        json file
+    """
+
+    def __init__(
+        self,
+        store_path: str,
+        podcast_downloads_path: str,
+        file_handler: StoreFileHandler,
+    ) -> None:
+        self._store_path = store_path
+        self._podcast_downloads_path = podcast_downloads_path
+
+        self._file_handler = file_handler
+
+        podcast_data = file_handler.read_data()
+        self.podcasts = StorePodcasts(
+            podcast_data=podcast_data, podcast_downloads_path=podcast_downloads_path
+        )
+
+    @classmethod
+    def create(
+        cls,
+        store_path: str,
+        store_file_path: str,
+        podcast_downloads_path: str,
+        setup_git: bool,
+        git_url: Optional[str] = None,
+        store_file_handler_cls: StoreFileHandler = UnencryptedStoreFileHandler,
+    ):
+        """Initialize a new pod store.
+
+        Optionally set up the `git` repo for the store.
+        """
+        try:
+            os.makedirs(store_path)
+        except FileExistsError:
+            raise StoreExistsError(store_path)
+        os.makedirs(podcast_downloads_path, exist_ok=True)
+
+        file_handler = store_file_handler_cls.create_with_file(store_file_path)
+
+        if setup_git:
+            run_git_command("init")
+            if git_url:
+                run_git_command(f"remote add origin {git_url}")
+
+        return cls(
+            store_path=store_path,
+            podcast_downloads_path=podcast_downloads_path,
+            file_handler=file_handler,
+        )
+
+    def __repr__(self):
+        return f"<Store({self._store_path!r})>"
+
+    def save(self):
+        """Save data to the store json file."""
+        podcast_data = self.podcasts.to_json()
+        self._file_handler.write_data(podcast_data)
